requirements.txt
asgiref==3.8.1
blinker==1.9.0
chardet==5.2.0
colorama==0.4.6
Django==5.1.3
et_xmlfile==2.0.0
greenlet==3.1.1
gunicorn==23.0.0
openpyxl==3.1.5
packaging==24.2
pillow==10.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
